refactor: Remove delete functionality from NetworkCard component; implement it in NetworkDetail component
