+++ client/src/pages/Index/Networks/index.tsx
@@ -77,8 +77,6 @@ const Networks = () => {
             <NetworkCard
               key={index}
               network={network}
-              networks={networks}
-              setNetworks={setNetworks}
             />
           );
         })}
